+++ eras/shelley/test-suite/test/Test/Cardano/Ledger/Shelley/Serialisation/Golden/Encoding.hs
@@ -518,6 +518,7 @@ tests =
                             { pmUrl = Maybe.fromJust $ textToUrl 64 poolUrl
                             , pmHash = poolMDHash
                             }
+                    , sppLeiosKey = SNothing
                     }
                 )
             )